+++ lib/roadie.rb
@@ -4,6 +4,8 @@ end
 require 'roadie/version'
 require 'roadie/errors'
 
+require 'roadie/utils'
+
 require 'roadie/stylesheet'
 require 'roadie/selector'
 require 'roadie/style_property'