+++ lib/roadie/asset_scanner.rb
@@ -80,7 +80,7 @@ module Roadie
     end
 
     def read_link_element(element)
-      if element['media'] != "print" && element["href"]
+      if element['media'] != "print" && element["href"] && !Utils.path_is_absolute?(element["href"])
         asset_provider.find_stylesheet! element['href']
       end
     end