+++ spec/lib/roadie/utils_spec.rb
@@ -1,0 +1,27 @@
+require "spec_helper"
+
+module Roadie
+  describe Utils, "path_is_absolute?" do
+    RSpec::Matchers.define :be_absolute do
+      match { |path| Utils.path_is_absolute?(path) }
+    end
+
+    it "detects absolute HTTP URLs" do
+      expect("http://example.com").to be_absolute
+      expect("https://example.com").to be_absolute
+      expect("https://example.com/path?foo=bar").to be_absolute
+    end
+
+    it "detects absolute URLs without schemes" do
+      expect("//example.com").to be_absolute
+      expect("//").to be_absolute
+    end
+
+    it "detects relative URLs without hosts" do
+      expect("path/to/me").not_to be_absolute
+      expect("/path/to/me").not_to be_absolute
+      expect("../../path").not_to be_absolute
+      expect("/").not_to be_absolute
+    end
+  end
+end